// legacy_ux/src/api/types/ontology.ts
export interface ListResponse<T> {
  items: T[];
  total: number;
  limit: number;
  offset: number;
}